# Disable forced inlining in arithmetic