I have a workload where I discovered that I got significantly better performance by deleting the `@inline` calls in the arithmetic functions. The workload is a little bit complicated, so I'll just show timing info. 

Using master:
```
julia> @time RegisterAffineAD.optimize_rigid(fixed, moving)

******************************************************************************
This program contains Ipopt, a library for large-scale nonlinear optimization.
 Ipopt is released as open source code under the Eclipse Public License (EPL).
         For more information visit http://projects.coin-or.org/Ipopt
******************************************************************************

  7.253977 seconds (138.38 M allocations: 7.238 GB, 20.80% gc time)
([-0.26290299117443555,-0.04562260395244045,0.09707317047564838],0.0007446994302309248)

julia> @time RegisterAffineAD.optimize_rigid(fixed, moving)
  5.894770 seconds (137.43 M allocations: 7.197 GB, 24.73% gc time)
([-0.26290299117443555,-0.04562260395244045,0.09707317047564838],0.0007446994302309248)

julia> @time RegisterAffineAD.optimize_rigid(fixed, moving)
  5.984184 seconds (137.43 M allocations: 7.197 GB, 24.45% gc time)
([-0.26290299117443555,-0.04562260395244045,0.09707317047564838],0.0007446994302309248)

julia> @time RegisterAffineAD.optimize_rigid(fixed, moving)
  5.896002 seconds (137.43 M allocations: 7.197 GB, 24.55% gc time)
([-0.26290299117443555,-0.04562260395244045,0.09707317047564838],0.0007446994302309248)
```

Using this PR:
```
julia> @time RegisterAffineAD.optimize_rigid(fixed, moving)

******************************************************************************
This program contains Ipopt, a library for large-scale nonlinear optimization.
 Ipopt is released as open source code under the Eclipse Public License (EPL).
         For more information visit http://projects.coin-or.org/Ipopt
******************************************************************************

  5.137296 seconds (969.36 k allocations: 1.095 GB, 2.14% gc time)
([-0.26290299117443555,-0.04562260395244045,0.09707317047564838],0.0007446994302309248)

julia> @time RegisterAffineAD.optimize_rigid(fixed, moving)
  4.066025 seconds (15.38 k allocations: 1.054 GB, 4.84% gc time)
([-0.26290299117443555,-0.04562260395244045,0.09707317047564838],0.0007446994302309248)

julia> @time RegisterAffineAD.optimize_rigid(fixed, moving)
  3.983676 seconds (14.87 k allocations: 1.054 GB, 1.79% gc time)
([-0.26290299117443555,-0.04562260395244045,0.09707317047564838],0.0007446994302309248)

julia> @time RegisterAffineAD.optimize_rigid(fixed, moving)
  4.142620 seconds (14.87 k allocations: 1.054 GB, 2.05% gc time)
([-0.26290299117443555,-0.04562260395244045,0.09707317047564838],0.0007446994302309248)
```

It seems there are 121 remaining uses of `@inline` in the code. I bet there are many other places where it might be a good idea to remove it. (I generally try to not use it unless I've checked that it actually improves performance; don't know whether that was done here or not.)