@mlubin 2015-09-26
Unfortunately the inlining instructions are necessary to get decent performance on many benchmarks. How are you using ForwardDiff? Chunk mode or vector mode?