@jrevels 2015-09-26
To elaborate, we had some benchmarks in which normal `Dual` numbers were beating out single-epsilon `GradientNumber`s for no discernible reason (they are the same theoretical construction); it turns out that inlining the arithmetic operations fixed it.

This incident prompted me to believe "if you expect the compiler to inline this, you should just manually inline it yourself as the inlining heuristics in Base can't always be relied upon." I then applied the `@inline` macro to a lot of places where the function body was "small." I'm now not necessarily certain this was the right strategy, and I was probably more heavy-handed in my approach than I should've been...

@timholy Would it be possible for you to link/post the code where ForwardDiff is being applied?