@KristofferC 2015-09-26
The number of allocations went down with a factor of 10 000 by turning off inlining? Is it because there is some type instability and the function call acts like a barrier?