@timholy 2015-09-26
@KristofferC, that was my first guess---I sometimes deliberately put `@noinline` on functions that act as a barrier---but I didn't yet have time to track it down. I'll see if I can post something.