@timholy 2015-09-27
I've posted a reduced test case [in this gist](https://gist.github.com/timholy/5d01015a18b5e9a4acdc). The code in `mygetindex` comes from `macroexpand`ing the `@generated` functions in Interpolations (I tested with the 0.3.0 release of Interpolations).

If I run with ForwardDiff master:
```jl
julia> include("perfFD.jl")
2-element Array{Float64,1}:
  123.305
 -952.366

julia> @time 1
  0.000006 seconds (148 allocations: 10.151 KB)
1

julia> @time sumA(itp.coefs, dz, irng, jrng)
  0.180838 seconds (5 allocations: 176 bytes)
2.4379931132108243e6

julia> @time gx, gy = gradsum(itp, dz, irng, jrng)
  0.223820 seconds (10 allocations: 368 bytes)
(123.30547906761386,-952.3663443995564)

julia> @time gg(dz)
  2.090166 seconds (87.67 M allocations: 2.613 GB, 21.22% gc time)
2-element Array{Float64,1}:
  123.305
 -952.366
```
If instead I use this PR:
```jl
julia> @time gg(dz)
  1.289657 seconds (9.74 M allocations: 297.263 MB, 3.88% gc time)
2-element Array{Float64,1}:
 -1072.05  
   -46.4687
```