@timholy 2015-09-27
Even more simplified test is the `perfFD2.jl` script in that same gist. It reduces the trouble to a single (complicated) line.