@mlubin 2015-09-27
@timholy, could you try passing the ``chunk_size`` argument to ``ForwardDiff.gradient``? That will likely have a significant improvement in performance.