@timholy 2015-09-27
I'm not exactly sure what `chunk_size` should be or what it does, but since my input is 2-dimension (i.e., `length(dz) = 2`) I tried `chunk_size=2`.

```jl
julia> @time gg(dz)
  2.042326 seconds (87.67 M allocations: 2.613 GB, 22.40% gc time)
2-element Array{Float64,1}:
  899.534
 -347.626

julia> gg = ForwardDiff.gradient(dz->sumA(itp.coefs, dz, irng, jrng), chunk_size=2);

julia> @time gg(dz)
  2.053191 seconds (87.67 M allocations: 2.613 GB, 22.30% gc time)
2-element Array{Float64,1}:
  899.534
 -347.626
```

No difference.