@timholy 2015-09-27
Oh, and I didn't explain the obvious (to me): `gradsum` corresponds to a hand-written gradient computation, and it's about 10x faster than ForwardDiff. However, with all the memory allocation, I think we're not yet to the point of finding out the performance limits of ForwardDiff.