@timholy 2015-09-27
Also, the dependency on Interpolations is only for comparison sake. That gist reduces the performance problem to 6 `GradientNumber`s, 6 integers, 9 `Float64`s, 12 multiplies, and 8 adds.