@jrevels 2015-09-27
I won't have time to really dig into this code until later (I need to pull METADATA to grab Interpolations.jl), so apologies if the answer here is obvious, but for this function:

```julia
function sumA(itp, dz, irng, jrng)
    s = 0.0
    dx, dy = dz[1], dz[2]
    for j in jrng
        for i in irng
            # s += itp[i+dx,j+dy]
            s += mygetindex(itp, i+dx, j+dy)
        end
    end
    s
end
```

Is the final type of `s` going to be a `GradientNumber` if `dz` holds `GradientNumber`s? If so, you're going to run into type instability/inferencing problems there due to initializing `s` as a float instead of saying something like `s=zero(eltype(dz))`.

P.S. There's documentation on chunk mode vs. vector mode [here](http://www.juliadiff.org/ForwardDiff.jl/chunk_vec_modes.html). One of the practical differences of the two is the use of tuples vs. vectors for partials storage.