@jrevels 2015-09-27
P.P.S Using chunk-mode here shouldn't change anything if your input dimension is only 2; ForwardDiff.jl defaults to tuple storage for low input dimensions anyway, so manually triggering chunk-mode won't have any different effect.