@timholy 2015-09-27
> Is the final type of s going to be a GradientNumber if dz holds GradientNumbers? If so, you're going to run into type instability/inferencing problems there due to initializing s as a float instead of saying something like s=zero(eltype(dz)).

Hmm, one would have thought that should have been obvious...

Helps, but still pretty slow. Master:
```jl
  1.698332 seconds (77.93 M allocations: 2.322 GB, 21.04% gc time)
```
This PR:
```jl
1.025838 seconds (130 allocations: 5.953 KB)
```
Number of allocations is way down, however!