@mlubin 2015-09-27
Would be interesting to the see the performance of dual numbers here.