@timholy 2015-09-27
```
julia> @time sumA(itp.coefs, dzd, irng, jrng)
dz = DualNumbers.Dual{Float64}[0.302642145419342 + 1.0du,0.9309871095886213 + 0.0du]
  0.231094 seconds (54 allocations: 2.344 KB)
2.4333389979457306e6 + 298.16242299983907du
```
That's just for one component, so I suppose you'd have to double it. But it's still faster than ForwardDiff.