@timholy 2015-09-27
If I take out my `@show dz` addition, then DualNumbers is non-allocating for this problem. So it's not a type-stability problem with my code.