@KristofferC 2015-09-27
Isn't `zero(eltype(dz))` wrong? `eltype` of a GradientNumber is just the type of it's real part, no?

Ex:

```jl
julia> nt
ForwardDiff.GradientNumber{2,Float64,Tuple{Float64,Float64}}

julia> zero(eltype(nt))
0.0

julia> zero(nt)
ForwardDiff.GradientNumber{2,Float64,Tuple{Float64,Float64}}(0.0,ForwardDiff.Partials{Float64,Tuple{Float64,Float64}}((0.0,0.0)))
```