@jrevels 2015-09-27
> Isn't zero(eltype(dz)) wrong? eltype of a GradientNumber is just the type of it's real part, no?

`dz` is a `Vector{G<:GradientNumber}`, unless I'm mistaken.