@jrevels 2015-09-28
Comparing this PR with master via [our feeble suite of benchmarks](https://github.com/JuliaDiff/ForwardDiff.jl/blob/master/benchmarks/benchmarks.jl):

Master:
```julia
julia> ack_bench = get_benchmark(ackley)
3x4 DataFrames.DataFrame
| Row | time        | func | xlen | chunk_size |
|-----|-------------|------|------|------------|
| 1   | 0.000118116 | 'f'  | 5000 | -1         |
| 2   | 0.867284    | 'g'  | 5000 | 1          |
| 3   | 0.208063    | 'g'  | 5000 | 5          |

julia> ros_bench = get_benchmark(rosenbrock)
3x4 DataFrames.DataFrame
| Row | time      | func | xlen | chunk_size |
|-----|-----------|------|------|------------|
| 1   | 1.244e-5  | 'f'  | 5000 | -1         |
| 2   | 0.0974855 | 'g'  | 5000 | 1          |
| 3   | 0.0636509 | 'g'  | 5000 | 5          |

julia> logit_bench = get_benchmark(self_weighted_logit)
3x4 DataFrames.DataFrame
| Row | time      | func | xlen | chunk_size |
|-----|-----------|------|------|------------|
| 1   | 6.689e-6  | 'f'  | 5000 | -1         |
| 2   | 0.131128  | 'g'  | 5000 | 1          |
| 3   | 0.0615632 | 'g'  | 5000 | 5          |
```

This PR:

```julia
julia> ack_bench = get_benchmark(ackley)
3x4 DataFrames.DataFrame
| Row | time        | func | xlen | chunk_size |
|-----|-------------|------|------|------------|
| 1   | 0.000119679 | 'f'  | 5000 | -1         |
| 2   | 1.15358     | 'g'  | 5000 | 1          |
| 3   | 0.297316    | 'g'  | 5000 | 5          |

julia> ros_bench = get_benchmark(rosenbrock)
3x4 DataFrames.DataFrame
| Row | time      | func | xlen | chunk_size |
|-----|-----------|------|------|------------|
| 1   | 1.2382e-5 | 'f'  | 5000 | -1         |
| 2   | 0.547275  | 'g'  | 5000 | 1          |
| 3   | 0.214978  | 'g'  | 5000 | 5          |

julia> logit_bench = get_benchmark(self_weighted_logit)
3x4 DataFrames.DataFrame
| Row | time      | func | xlen | chunk_size |
|-----|-----------|------|------|------------|
| 1   | 6.736e-6  | 'f'  | 5000 | -1         |
| 2   | 0.221926  | 'g'  | 5000 | 1          |
| 3   | 0.0869565 | 'g'  | 5000 | 5          |
```

Under the `func` column, `'f'` denotes the original function evaluation time, while `'g'` denotes the gradient evaluation time.

I swear, the next big addition to ForwardDiff will be a real `ForwardDiffBenchmarks` module...with docs.